refactor: Move `ModifierKeyCode` to new module
