+++ src/event/key_modifier.rs
@@ -1,0 +1,97 @@
+use super::*;
+
+/// Represents a modifier key (as part of [`KeyCode::Modifier`]).
+#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Clone, Copy, Hash)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub enum ModifierKeyCode {
+    /// Left Shift key.
+    LeftShift,
+    /// Left Control key. (Control on macOS, Ctrl on other platforms)
+    LeftControl,
+    /// Left Alt key. (Option on macOS, Alt on other platforms)
+    LeftAlt,
+    /// Left Super key. (Command on macOS, Windows on Windows, Super on other platforms)
+    LeftSuper,
+    /// Left Hyper key.
+    LeftHyper,
+    /// Left Meta key.
+    LeftMeta,
+    /// Right Shift key.
+    RightShift,
+    /// Right Control key. (Control on macOS, Ctrl on other platforms)
+    RightControl,
+    /// Right Alt key. (Option on macOS, Alt on other platforms)
+    RightAlt,
+    /// Right Super key. (Command on macOS, Windows on Windows, Super on other platforms)
+    RightSuper,
+    /// Right Hyper key.
+    RightHyper,
+    /// Right Meta key.
+    RightMeta,
+    /// Iso Level3 Shift key.
+    IsoLevel3Shift,
+    /// Iso Level5 Shift key.
+    IsoLevel5Shift,
+}
+
+impl Display for ModifierKeyCode {
+    /// Formats the modifier key using the given formatter.
+    ///
+    /// # Platform-specific Notes
+    ///
+    /// On macOS, the control, alt, and super keys are displayed as "Control", "Option", and
+    /// "Command" respectively. See
+    /// <https://support.apple.com/guide/applestyleguide/welcome/1.0/web>.
+    ///
+    /// On Windows, the super key is displayed as "Windows" and the control key is displayed as
+    /// "Ctrl". See
+    /// <https://learn.microsoft.com/en-us/style-guide/a-z-word-list-term-collections/term-collections/keys-keyboard-shortcuts>.
+    ///
+    /// On other platforms, the super key is referred to as "Super".
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ModifierKeyCode::LeftShift => write!(f, "Left Shift"),
+            ModifierKeyCode::LeftHyper => write!(f, "Left Hyper"),
+            ModifierKeyCode::LeftMeta => write!(f, "Left Meta"),
+            ModifierKeyCode::RightShift => write!(f, "Right Shift"),
+            ModifierKeyCode::RightHyper => write!(f, "Right Hyper"),
+            ModifierKeyCode::RightMeta => write!(f, "Right Meta"),
+            ModifierKeyCode::IsoLevel3Shift => write!(f, "Iso Level 3 Shift"),
+            ModifierKeyCode::IsoLevel5Shift => write!(f, "Iso Level 5 Shift"),
+
+            #[cfg(target_os = "macos")]
+            ModifierKeyCode::LeftControl => write!(f, "Left Control"),
+            #[cfg(not(target_os = "macos"))]
+            ModifierKeyCode::LeftControl => write!(f, "Left Ctrl"),
+
+            #[cfg(target_os = "macos")]
+            ModifierKeyCode::LeftAlt => write!(f, "Left Option"),
+            #[cfg(not(target_os = "macos"))]
+            ModifierKeyCode::LeftAlt => write!(f, "Left Alt"),
+
+            #[cfg(target_os = "macos")]
+            ModifierKeyCode::LeftSuper => write!(f, "Left Command"),
+            #[cfg(target_os = "windows")]
+            ModifierKeyCode::LeftSuper => write!(f, "Left Windows"),
+            #[cfg(not(any(target_os = "macos", target_os = "windows")))]
+            ModifierKeyCode::LeftSuper => write!(f, "Left Super"),
+
+            #[cfg(target_os = "macos")]
+            ModifierKeyCode::RightControl => write!(f, "Right Control"),
+            #[cfg(not(target_os = "macos"))]
+            ModifierKeyCode::RightControl => write!(f, "Right Ctrl"),
+
+            #[cfg(target_os = "macos")]
+            ModifierKeyCode::RightAlt => write!(f, "Right Option"),
+            #[cfg(not(target_os = "macos"))]
+            ModifierKeyCode::RightAlt => write!(f, "Right Alt"),
+
+            #[cfg(target_os = "macos")]
+            ModifierKeyCode::RightSuper => write!(f, "Right Command"),
+            #[cfg(target_os = "windows")]
+            ModifierKeyCode::RightSuper => write!(f, "Right Windows"),
+            #[cfg(not(any(target_os = "macos", target_os = "windows")))]
+            ModifierKeyCode::RightSuper => write!(f, "Right Super"),
+        }
+    }
+}
